+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   `reflect.Kind`.
 - Skip populating function and value types instead of reporting errors.
 - **[Breaking]** `Value.Timestamp` is private, use Value.LastUpdated instead.
+- **[Breaking]** Most of the `Provider` constructors return an error instead of panics.
+- **[Breaking]** `Value.WithDefault` returns an error when a default can't be used.
 
 ## v1.0.0-rc1 (26 Jun 2017)